+++ lib/money-rails/active_record/monetizable.rb
@@ -84,33 +84,33 @@ module MoneyRails
             # All the options which are available for Rails numericality
             # validation, are also available for both types.
             # E.g.
-            #   monetize :price_in_a_range_cents, :allow_nil => true,
-            #     :subunit_numericality => {
-            #       :greater_than_or_equal_to => 0,
-            #       :less_than_or_equal_to => 10000,
+            #   monetize :price_in_a_range_cents, allow_nil: true,
+            #     subunit_numericality: {
+            #       greater_than_or_equal_to: 0,
+            #       less_than_or_equal_to: 10000,
             #     },
-            #     :numericality => {
-            #       :greater_than_or_equal_to => 0,
-            #       :less_than_or_equal_to => 100,
-            #       :message => "must be greater than zero and less than $100"
+            #     numericality: {
+            #       greater_than_or_equal_to: 0,
+            #       less_than_or_equal_to: 100,
+            #       message: "must be greater than zero and less than $100"
             #     }
             #
             # To disable validation entirely, use :disable_validation, E.g:
-            #   monetize :price_in_a_range_cents, :disable_validation => true
+            #   monetize :price_in_a_range_cents, disable_validation: true
             if (validation_enabled = MoneyRails.include_validations && !options[:disable_validation])
 
               # This is a validation for the subunit
               if (subunit_numericality = options.fetch(:subunit_numericality, true))
                 validates subunit_name, {
-                  :allow_nil => options[:allow_nil],
-                  :numericality => subunit_numericality
+                  allow_nil: options[:allow_nil],
+                  numericality: subunit_numericality
                 }
               end
 
               # Allow only Money objects or Numeric values!
               if (numericality = options.fetch(:numericality, true))
                 validates name.to_sym, {
-                  :allow_nil => options[:allow_nil],
+                  allow_nil: options[:allow_nil],
                   'money_rails/active_model/money' => numericality
                 }
               end
@@ -119,7 +119,7 @@ module MoneyRails
 
             # Getter for monetized attribute
             define_method name do |*args|
-              read_monetized name, subunit_name, *args
+              read_monetized name, subunit_name, options, *args
             end
 
             # Setter for monetized attribute
@@ -178,10 +178,11 @@ module MoneyRails
         end
       end
 
-      def read_monetized(name, subunit_name, *args)
+      def read_monetized(name, subunit_name, options = {}, *args)
         # Get the cents
         amount = public_send(subunit_name, *args)
 
+        return if amount.nil? && options[:allow_nil]
         # Get the currency object
         attr_currency = public_send("currency_for_#{name}")