+++ lib/money-rails/mongoid/money.rb
@@ -3,8 +3,8 @@ class Money
   # Converts an object of this instance into a database friendly value.
   def mongoize
     {
-      :cents        => cents.mongoize.to_f,
-      :currency_iso => currency.iso_code.mongoize
+      cents:        cents.mongoize.to_f,
+      currency_iso: currency.iso_code.mongoize
     }
   end