+++ spec/dummy/db/migrate/20120331190108_create_products.rb
@@ -1,4 +1,4 @@
-class CreateProducts < ActiveRecord::Migration
+class CreateProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
   def change
     create_table :products do |t|
       t.integer :price_cents