+++ spec/dummy/db/migrate/20120528210103_create_dummy_products.rb
@@ -1,4 +1,4 @@
-class CreateDummyProducts < ActiveRecord::Migration
+class CreateDummyProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
   def change
     create_table :dummy_products do |t|
       t.string :currency