+++ spec/dummy/db/migrate/20141005075025_add_aliased_attr_to_products.rb
@@ -1,4 +1,4 @@
-class AddAliasedAttrToProducts < ActiveRecord::Migration
+class AddAliasedAttrToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
   def change
     add_column :products, :aliased_cents, :integer
   end