+++ spec/dummy/db/migrate/20150126231442_add_reduced_price_to_products.rb
@@ -1,4 +1,4 @@
-class AddReducedPriceToProducts < ActiveRecord::Migration
+class AddReducedPriceToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
   def change
     add_column :products, :reduced_price_cents, :integer
     add_column :products, :reduced_price_currency, :string