+++ spec/dummy/db/migrate/20150217222612_add_lambda_price_to_products.rb
@@ -1,4 +1,4 @@
-class AddLambdaPriceToProducts < ActiveRecord::Migration
+class AddLambdaPriceToProducts < (Rails::VERSION::MAJOR >= 5 ? ActiveRecord::Migration[4.2] : ActiveRecord::Migration)
   def change
     add_column :products, :lambda_price_cents, :integer
   end